// src/interfaces/LocalePermission.ts
// Generated by https://quicktype.io

export interface ILocalePermission {
  id: number;
  attributes: LocalePermissionAttributes;
}

export interface LocalePermissionAttributes {
  cep: string;
  cityIBGECod: number;
  cityName: string;
  stateName: string;
  stateIBGECod: number;
  cityPermisson: boolean;
  statePermission: boolean;
  active: boolean;
  createdAt: string;
  updatedAt: string;
  publishedAt: string;
}